Adding new send method in producer trait to allow to send direct producer messages
Adding helpers to search for headers
Removing not needed execution contexts
Setting scheduler as part of the main execution
Making the compiler happier
Upgrading Monix version
Removing deprecated method in commit logic
Upgrading scala version to 2.12.8
Adding some helpers to process consumer records: sync, task, obs

+++ src/main/scala/com/ubirch/kafka/consumer/WithConsumerShutdownHook.scala
@@ -9,7 +9,7 @@ trait WithConsumerShutdownHook extends LazyLogging {
     () =>
       logger.info(s"Shutting down Consumer[timeout=$gracefulTimeout secs name=${consumerRunner.getName}] ...")
       if (Option(consumerRunner).isDefined)
-        Future.successful(consumerRunner.shutdown(gracefulTimeout, java.util.concurrent.TimeUnit.SECONDS))
+        Future.successful(consumerRunner.shutdown(gracefulTimeout.toLong, java.util.concurrent.TimeUnit.SECONDS))
       else Future.unit
   }
 }
